removed short consts

+++ src/vate/org/vate/VT.java
@@ -21,12 +21,12 @@ public class VT
 	public static final int VT_DATA_BUFFER_SIZE = 1024 * 32;
 	public static final int VT_SMALL_BUFFER_SIZE = 1024 * 8;
 	
-	public static final short VT_MULTIPLEXED_CHANNEL_TYPE_PIPED = 0;
-	public static final short VT_MULTIPLEXED_CHANNEL_TYPE_DIRECT = 1;
-	public static final short VT_MULTIPLEXED_CHANNEL_TYPE_PERFORMANCE_THROTTLEABLE = 0 << 1;
-	public static final short VT_MULTIPLEXED_CHANNEL_TYPE_PERFORMANCE_UNLIMITED = 1 << 1;
-	public static final short VT_MULTIPLEXED_CHANNEL_TYPE_COMPRESSION_DISABLED = 0 << 2;
-	public static final short VT_MULTIPLEXED_CHANNEL_TYPE_COMPRESSION_ENABLED = 1 << 2;
+	public static final int VT_MULTIPLEXED_CHANNEL_TYPE_PIPED = 0;
+	public static final int VT_MULTIPLEXED_CHANNEL_TYPE_DIRECT = 1;
+	public static final int VT_MULTIPLEXED_CHANNEL_TYPE_PERFORMANCE_THROTTLEABLE = 0 << 1;
+	public static final int VT_MULTIPLEXED_CHANNEL_TYPE_PERFORMANCE_UNLIMITED = 1 << 1;
+	public static final int VT_MULTIPLEXED_CHANNEL_TYPE_COMPRESSION_DISABLED = 0 << 2;
+	public static final int VT_MULTIPLEXED_CHANNEL_TYPE_COMPRESSION_ENABLED = 1 << 2;
 	//public static final short VT_MULTIPLEXED_CHANNEL_TYPE_COMPRESSION_LZ4 = 0 << 3;
 	//public static final short VT_MULTIPLEXED_CHANNEL_TYPE_COMPRESSION_DEFLATE = 1 << 3;
